Move QuestionDataSelector and its module to new folder and convert to CSS modules
