+++ frontend/src/metabase/query_builder/components/view/QuestionDataSelector/QuestionDataSelector.module.css
@@ -1,4 +1,4 @@
 /* HACK: DataPopover should be below the search box */
-:global(.DataPopoverContainer) {
+.DataPopoverContainer {
   z-index: 2;
 }
